Create slugify mixin and test with weekday model
